+++ CHANGES.rst
@@ -18,6 +18,8 @@ History
 
 - Added 'delete' command.
 
+- Add metadata to schematisation output.
+
 
 0.1.0 (2021-11-11)
 ------------------